requirements.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
googletrans==4.0.0rc1
